# api/user.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/user.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "User"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/shellverse-service/user/login": {
      "post": {
        "operationId": "User_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiUserLoginRequest"
            }
          }
        ],
        "tags": [
          "User"
        ]
      }
    }
  },
  "definitions": {
    "apiUserLoginRequest": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "title": "@gotags: binding:\"required\""
        },
        "userinfo": {
          "$ref": "#/definitions/apiUserLoginRequestUserInfo",
          "title": "@gotags: binding:\"required\""
        }
      }
    },
    "apiUserLoginRequestUserInfo": {
      "type": "object",
      "properties": {
        "nickname": {
          "type": "string",
          "title": "@gotags: binding:\"required\""
        },
        "avatar_url": {
          "type": "string",
          "title": "@gotags: binding:\"required\""
        }
      }
    },
    "apiUserLoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "userinfo": {
          "$ref": "#/definitions/apiUserLoginResponseUserInfo"
        }
      }
    },
    "apiUserLoginResponseUserInfo": {
      "type": "object",
      "properties": {
        "nickname": {
          "type": "string"
        },
        "avatar_url": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}